Implement front-end for grad-expressions
This includes:
- Lexing of the `grad`-keyword
- Parsing of `GradExpr`
- Type-checking and infering

Code generation will be added later.

Currently grad expressions only work for functions that:
- take at least one float
- return exactly one float

Later on this will be extended for all input-types that can be differentiated.
Maybe the restriction on scalar return values will be lifted.

Currently the grad-expression maps functions of type
    I1 x I2 x ... x In -> O1
to functions of type
    I1 x I2 x ... x In -> I1 x I2 x ... x In

Later on this might be changed so that the tangent-vectors of the input-types
are returned. In most cases these tangent vectors are the input-type itself,
but this is not always true (see https://github.com/dan-zheng/swift/blob/differentiable-programming/docs/DifferentiableProgramming.md#the-differentiable-protocol).