Add tangent-vector types for most types
Now types like arrays of floats or tuples can be used as parameters for differentiable functions.
The elements of composite type tangent-vectors are the tangent-vectors of the composite type's elements.

If a type's elements have no tangent vectors then unit is returned.
Note:
- This is not yet completely consistent for tuples.
- This is not yet implemented for tuples.

Some examples:

    [f32 * 4] -> [f32 * 4]
    (i32, i32, f32) -> ((), (), f32)
    [(i32, f32)] -> [((), f32)]

Also I removed the check that assured at least some of the input parameters have a tangent type.
I think this is not needed as the return type `unit` should be enough.